Change quiz definition
With the previous QuizDefinition, it was not possible to have multiple answers
and multiple alphabets displayed in the question field, which was undesired in
situations like Japanese, where kanji does not indicate how it is read.

This path updates the database schema and the Question and QuizDefinition
activities to fit the requirements.